expose render_to
this exposes `render_to`, and moves the call to `present()` to
`render_async` instead. this way, `render_to` can return the
`SubmissionIndex` and non-web targets can use `device.poll()` to wait
for the frame at later time, presenting it then.

my render loop calls `render_to` on a timer to limit fps and stores the
`SubmissionIndex` until a Wayland frame callback comes in and i need to
actually `.present()`.