Improves workflow for the native app in pause
- Set the pause state in the title bar
it's simple but sometime it's not clear if the app is running or not and
when the laptop is on battery it's nice to know the state of the app.

- Force a redraw when a change is done in the file
When working with the application paused, it's still useful to get update
of the window when a change is done in the file.

Before this commit the window was not always updated when a change in
the file was detected. With this commit we force an event in the
main_event_loop to trigger the redraw.